requirements.txt
Flask~=2.2.2

redis~=4.5.1
mysql-connector-python~=8.0.32
PyJWT~=2.6.0
requests~=2.28.2
urllib3~=1.26.14
jproperties~=2.1.1